# contracts/openapi/public-api.yaml
openapi: 3.1.0
info:
  title: AnchorPay Public API
  version: 1.0.0
  summary: Customer, staff and webhook API exposed through the API gateway.
  description: |
    All client traffic enters through the API gateway, which validates the JWT, applies rate limits,
    adds `X-Request-Id` and routes each path to the service named in `x-owner-service`.

    **Conventions** (frozen at Foundation sign-off)
    - JSON bodies use camelCase. Money is `{ amountMinor, currency }` in integer minor units; rates are decimal strings.
    - Every error is `application/problem+json` with a stable `code` (see `ErrorCode`).
    - `x-roles` lists the roles allowed to call an operation. Customers only ever see their own resources.
    - Versioning: breaking changes go to `/v2`; additive changes (new optional fields, new endpoints) stay in `/v1`.
    - Access tokens expire after 15 minutes; refresh tokens after 30 days (rotated on every refresh).
  contact:
    name: AnchorPay engineering
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://127.0.0.1:8080
    description: Local API gateway
security:
  - bearerAuth: []
...
paths:
...
  /v1/users:
    post:
      tags: [Auth]
      operationId: registerUser
      summary: Register a new customer account
      description: >-
        Creates a customer at KYC tier 0 and sends an email verification link and a phone code.
        Only Canadian residents can register while the send side is Canada-only.
      security: []
      x-owner-service: identity-service
      x-roles: [public]
      parameters:
        - $ref: 'common.yaml#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: 'common.yaml#/components/responses/BadRequest'
        '409':
          $ref: 'common.yaml#/components/responses/Conflict'
        '429':
          $ref: 'common.yaml#/components/responses/TooManyRequests'
        default:
          $ref: 'common.yaml#/components/responses/ServerError'
...
  /v1/auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      summary: Exchange a refresh token for a new token pair
      description: The old refresh token is revoked (rotation). Re-using a revoked token revokes the whole session.
      security: []
      x-owner-service: identity-service
      x-roles: [public]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400':
          $ref: 'common.yaml#/components/responses/BadRequest'
        '401':
          $ref: 'common.yaml#/components/responses/Unauthorized'
        default:
          $ref: 'common.yaml#/components/responses/ServerError'
...
  /v1/auth/verify-email:
    post:
      tags: [Auth]
      operationId: verifyEmail
      summary: Confirm the email address with the token from the verification link
      security: []
      x-owner-service: identity-service
      x-roles: [public]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '204':
          description: Email verified
        '400':
          $ref: 'common.yaml#/components/responses/BadRequest'
        '404':
          $ref: 'common.yaml#/components/responses/NotFound'
        default:
          $ref: 'common.yaml#/components/responses/ServerError'
  /v1/auth/verify-phone:
    post:
      tags: [Auth]
      operationId: verifyPhone
      summary: Confirm the phone number with the 6-digit SMS code
      x-owner-service: identity-service
      x-roles: [customer]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  pattern: '^[0-9]{6}$'
      responses:
        '204':
          description: Phone verified
        '400':
          $ref: 'common.yaml#/components/responses/BadRequest'
        '401':
          $ref: 'common.yaml#/components/responses/Unauthorized'
        '429':
          $ref: 'common.yaml#/components/responses/TooManyRequests'
        default:
          $ref: 'common.yaml#/components/responses/ServerError'
  /v1/auth/resend-verification:
    post:
      tags: [Auth]
      operationId: resendVerification
      summary: Resend the email link or SMS code
      x-owner-service: identity-service
      x-roles: [customer]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel]
              properties:
                channel:
                  type: string
                  enum: [email, phone]
      responses:
        '204':
          description: Sent (limited to 1 per minute, 5 per day)
        '401':
          $ref: 'common.yaml#/components/responses/Unauthorized'
        '429':
          $ref: 'common.yaml#/components/responses/TooManyRequests'
        default:
          $ref: 'common.yaml#/components/responses/ServerError'
  /v1/auth/forgot-password:
    post:
      tags: [Auth]
      operationId: forgotPassword
      summary: Email a password reset link
      description: Always returns 204, whether or not the email exists, so accounts can't be discovered.
      security: []
      x-owner-service: identity-service
      x-roles: [public]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '204':
          description: Request accepted
        '400':
          $ref: 'common.yaml#/components/responses/BadRequest'
        '429':
          $ref: 'common.yaml#/components/responses/TooManyRequests'
        default:
          $ref: 'common.yaml#/components/responses/ServerError'
  /v1/auth/reset-password:
    post:
      tags: [Auth]
      operationId: resetPassword
      summary: Set a new password using the reset token (valid 30 minutes, single use)
      security: []
      x-owner-service: identity-service
      x-roles: [public]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token:
                  type: string
                newPassword:
                  $ref: '#/components/schemas/Password'
      responses:
        '204':
          description: Password changed; all sessions revoked
        '400':
          $ref: 'common.yaml#/components/responses/BadRequest'
        '404':
          $ref: 'common.yaml#/components/responses/NotFound'
        default:
          $ref: 'common.yaml#/components/responses/ServerError'

  # ---------------------------------------------------------------- profile
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'RS256 access token from /v1/auth/login. Claims: sub (user id), role, sid (session id), iss, aud, exp (15 min). KYC tier is not a claim: it can change at any time, ask compliance-service.'
    stripeSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
    providerSignature:
      type: apiKey
      in: header
      name: X-Signature
      description: 'Hex HMAC-SHA256 of "<X-Timestamp>.<raw body>" with the provider webhook secret. Rejected if older than 5 minutes.'
...
  schemas:
...
    Password:
      type: string
      minLength: 12
      maxLength: 128
      writeOnly: true
      description: At least 12 characters. Checked against a list of common/breached passwords.
    Address:
      type: object
      required: [line1, city, province, postalCode]
      properties:
        line1:
          type: string
          maxLength: 200
        line2:
          type: string
          maxLength: 200
        city:
          type: string
          maxLength: 100
        province:
          type: string
          enum: [AB, BC, MB, NB, NL, NS, NT, NU, 'ON', PE, QC, SK, YT]
        postalCode:
          type: string
          pattern: '^[A-Za-z][0-9][A-Za-z] ?[0-9][A-Za-z][0-9]$'
    RegisterRequest:
      type: object
      required: [email, phone, password, fullName, dateOfBirth, country, acceptTerms]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        phone:
          type: string
          description: E.164 format.
          pattern: '^\+[1-9][0-9]{7,14}$'
          examples: ['+14165550123']
        password:
          $ref: '#/components/schemas/Password'
        fullName:
          type: string
          minLength: 2
          maxLength: 150
          description: As shown on government ID.
        dateOfBirth:
          type: string
          format: date
          description: Must be 18 or older.
        address:
          $ref: '#/components/schemas/Address'
        country:
          type: string
          const: CA
        acceptTerms:
          type: boolean
          const: true
...
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    TokenPair:
      type: object
      required: [accessToken, refreshToken, tokenType, expiresIn, refreshExpiresIn]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          const: Bearer
        expiresIn:
          type: integer
          description: Seconds (900).
        refreshExpiresIn:
          type: integer
          description: Seconds (2592000).
    User:
      type: object
      required: [id, email, phone, fullName, country, role, status, emailVerified, phoneVerified, createdAt]
      properties:
        id:
          $ref: 'common.yaml#/components/schemas/Uuid'
        email:
          type: string
          format: email
        phone:
          type: string
        fullName:
          type: string
        dateOfBirth:
          type: string
          format: date
        address:
          $ref: '#/components/schemas/Address'
        country:
          $ref: 'common.yaml#/components/schemas/Country'
        role:
          $ref: 'common.yaml#/components/schemas/Role'
        status:
          type: string
          enum: [active, locked, suspended, closed]
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
        createdAt:
          $ref: 'common.yaml#/components/schemas/Timestamp'
...